+++ TauriApiWrapper/Objects/Responses/Achievement/CharacterAchievementsInfo.cs
@@ -5,15 +5,15 @@ namespace TauriApiWrapper.Objects.Responses.Achievement
     public partial class CharacterAchievementsInfo
     {
         [JsonProperty("id")]
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         [JsonProperty("pid")]
-        public long Pid { get; set; }
+        public int Pid { get; set; }
 
         [JsonProperty("name")]
         public string Name { get; set; }
 
         [JsonProperty("order")]
-        public long Order { get; set; }
+        public int Order { get; set; }
     }
 }
